+++ app/Main.hs
@@ -2,6 +2,9 @@ module Main where
 
 import Snap
 import Site
+import EnvRead
 
 main :: IO ()
-main = serveSnaplet defaultConfig appInit
+main = do
+  ci <- loadConnInfoFromEnv
+  serveSnaplet defaultConfig $ appInit ci