+++ src/User/Queries.hs
@@ -1,0 +1,11 @@
+module User.Queries where
+
+import Database.PostgreSQL.Simple.SqlQQ
+import Snap.Snaplet.PostgresqlSimple
+
+allUsersQuery :: Query
+allUsersQuery = 
+  [sql| 
+    SELECT *
+    FROM snap_user 
+  |]